+++ Scenes/Menu_UI/debug_ui.tscn
@@ -1,46 +1,66 @@
 [gd_scene load_steps=4 format=3 uid="uid://ppqc7eiqvsaa"]
 
-[ext_resource type="Script" path="res://Scenes/Menu_UI/DebugUi.cs" id="1_s2phv"]
+[ext_resource type="Script" path="res://Scripts/UI_Menu/DebugUi.cs" id="1_s2phv"]
 [ext_resource type="Texture2D" uid="uid://baiy23homkmky" path="res://Assets/Menu_UI/arrow-keys.png" id="1_unmnx"]
 [ext_resource type="Texture2D" uid="uid://c7y64gmgbqkqy" path="res://Assets/Menu_UI/WASD.png" id="3_pkarl"]
 
-[node name="Debug UI" type="Control"]
-layout_mode = 3
+[node name="Debug UI" type="CanvasLayer" node_paths=PackedStringArray("highlight", "wasd", "arrowKeys")]
+script = ExtResource("1_s2phv")
+UP_POS = Vector2i(110, 90)
+highlight = NodePath("MarginContainer/Rows/Bot Row/ColorRect")
+wasd = NodePath("MarginContainer/Rows/Bot Row/Input Display/WASD")
+arrowKeys = NodePath("MarginContainer/Rows/Bot Row/Input Display/Arrow Keys")
+
+[node name="MarginContainer" type="MarginContainer" parent="."]
 anchors_preset = 15
 anchor_right = 1.0
 anchor_bottom = 1.0
 grow_horizontal = 2
 grow_vertical = 2
-script = ExtResource("1_s2phv")
+theme_override_constants/margin_left = 40
+theme_override_constants/margin_top = 40
+theme_override_constants/margin_right = 40
+theme_override_constants/margin_bottom = 40
 
-[node name="ColorRect" type="ColorRect" parent="."]
-visible = false
-layout_mode = 0
-offset_left = 856.0
-offset_top = 544.0
-offset_right = 944.0
-offset_bottom = 632.0
-color = Color(1, 1, 0, 1)
+[node name="Rows" type="VBoxContainer" parent="MarginContainer"]
+layout_mode = 2
 
-[node name="Arrow Keys" type="TextureRect" parent="."]
-layout_mode = 0
-offset_left = 844.0
-offset_top = 441.0
-offset_right = 1455.0
-offset_bottom = 849.0
-scale = Vector2(0.5, 0.5)
-size_flags_horizontal = 8
-size_flags_vertical = 8
+[node name="Top Row" type="HBoxContainer" parent="MarginContainer/Rows"]
+layout_mode = 2
+
+[node name="Mid Row" type="HBoxContainer" parent="MarginContainer/Rows"]
+layout_mode = 2
+size_flags_vertical = 3
+size_flags_stretch_ratio = 3.0
+
+[node name="Bot Row" type="HBoxContainer" parent="MarginContainer/Rows"]
+layout_mode = 2
+size_flags_vertical = 3
+
+[node name="Input Display" type="AspectRatioContainer" parent="MarginContainer/Rows/Bot Row"]
+layout_mode = 2
+size_flags_horizontal = 10
+
+[node name="Arrow Keys" type="TextureRect" parent="MarginContainer/Rows/Bot Row/Input Display"]
+custom_minimum_size = Vector2(300, 200)
+layout_mode = 2
+size_flags_horizontal = 0
+size_flags_vertical = 0
 texture = ExtResource("1_unmnx")
+expand_mode = 1
 
-[node name="WASD" type="TextureRect" parent="."]
+[node name="WASD" type="TextureRect" parent="MarginContainer/Rows/Bot Row/Input Display"]
 visible = false
-layout_mode = 0
-offset_left = 844.0
-offset_top = 441.0
-offset_right = 1455.0
-offset_bottom = 849.0
-scale = Vector2(0.5, 0.5)
-size_flags_horizontal = 8
-size_flags_vertical = 8
+custom_minimum_size = Vector2(300, 200)
+layout_mode = 2
+size_flags_horizontal = 0
+size_flags_vertical = 0
 texture = ExtResource("3_pkarl")
+expand_mode = 1
+
+[node name="ColorRect" type="ColorRect" parent="MarginContainer/Rows/Bot Row"]
+custom_minimum_size = Vector2(80, 80)
+layout_mode = 2
+size_flags_horizontal = 0
+size_flags_vertical = 0
+color = Color(1, 1, 0, 1)
